requirements.txt
streamlit
opencv-python
numpy
Pillow
ultralytics
supervision
